% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_bound.R
\name{selection_bound}
\alias{selection_bound}
\title{Sensitivity analysis for selection bias}
\usage{
selection_bound(
  y,
  x,
  w,
  z = NULL,
  L0l,
  L0u,
  L1,
  cons = NULL,
  theta = NULL,
  alpha = 0.05,
  opts = NULL
)
}
\arguments{
\item{y}{Outcome (vector)}

\item{x}{Explanatory variables (matrix)}

\item{w}{Selection variables (matrix)}

\item{z}{Optional instrumental variables (matrix)}

\item{L0l}{Lower bound for the probability of sample selection for the average observation}

\item{L0u}{Upper bound for the probability of sample selection for the average observation}

\item{L1}{Odds ratio of sample selection for a one unit (binary) or one standard
deviation (continuous) increase in a variable.}

\item{cons}{List of constraints to be applied: \code{RESP} is a response rate
constraint; \code{COVMEAN} is a covariate mean constraint; \code{DIREC} is a
directionality constraint.}

\item{theta}{Optional starting parameter for the global optimiser.}

\item{alpha}{Significance level for confidence interval.}

\item{opts}{Optional list of options for the global optimiser.}
}
\value{
Named list of objects: \code{theta_min} (\code{theta_max}) is the parameters
for the lower (upper) bound; \code{interval} is a vector containing the lower and upper
bounds; \code{ci} is a vector containing the \code{alpha}-level confidence interval.
}
\description{
Given a set of sensitivity parameters and constraints, computes an
upper and lower bound for an inverse probability weighted regression estimate
}
\examples{

}
